requirements.txt
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3
cryptography==3.1.1
Django==3.1.1
djangorestframework==3.11.1
future==0.18.2
idna==2.10
iniconfig==1.0.1
Jinja2==2.11.2
junos-eznc==2.5.3
lxml==4.5.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
ncclient==0.6.9
netaddr==0.8.0
ntc-templates==1.5.0
packaging==20.4
paramiko==2.7.2
pluggy==0.13.1
ply==3.11
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.8
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
pysmi==0.3.4
pysnmp==4.4.12
pytest==6.1.0
pytest-cov==2.10.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
scp==0.13.2
six==1.15.0
sqlparse==0.3.1
textfsm==1.1.0
toml==0.10.1
transitions==0.8.3
urllib3==1.25.10
yamlordereddictloader==0.4.0
